refactor(dashboard): move around components into appropriate folders
