--- packages/frontend/src/components/dashboard/home/ViewClassroom/index.ts
@@ -1,3 +1,0 @@
-import ViewClassroom from './ViewClassroom';
-
-export default ViewClassroom;
