--- packages/frontend/src/components/dashboard/home/calendars/index.ts
@@ -1,4 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-export { default as TeacherCalendar } from './TeacherCalendar';
-// TODO: add StudentCalendar component
-// export { default as StudentCalendar } from './StudentCalendar';
