--- packages/frontend/src/components/dashboard/home/forms/index.ts
@@ -1,2 +1,0 @@
-export { default as ClassroomCreation } from './ClassroomCreation';
-export { default as NewStudent } from './NewStudent';
